undo ddtrace and move fixtures back
