+++ ddtrace/contrib/pytest_benchmark/plugin.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ddtrace.contrib.pytest.plugin import _extract_span
+from ddtrace.contrib.pytest.plugin_v1 import _extract_span
 from ddtrace.contrib.pytest_benchmark.constants import BENCHMARK_INFO
 from ddtrace.contrib.pytest_benchmark.constants import PLUGIN_METRICS
 from ddtrace.contrib.pytest_benchmark.constants import PLUGIN_OUTLIERS
